refactor: Simplify product module validation and error handling

+++ src/modules/products/product.service.ts
@@ -1,4 +1,4 @@
-import { CreateProductDTO } from "./product.dto";
+import { ProductDTO } from "./product.dto";
 import { ProductRepository } from "./product.repository";
 import { deleteImageFromFirebase } from '../../services/firebase';
 import { ProductEntity } from '../../entities/product.entity';
@@ -6,13 +6,13 @@ import { ProductEntity } from '../../entities/product.entity';
 export class ProductService {
   constructor(private productRepository: ProductRepository = new ProductRepository()) { }
 
-  async createProduct(productDto: CreateProductDTO) {
+  async createProduct(productDto: ProductDTO) {
     const codeAlredyExists = await this.productRepository.getProductByCodigo(productDto.codigo_produto);
     if (codeAlredyExists) {
-      throw new Error("PRODUCT_DUPLICATE");
+      throw new Error("Produto com o mesmo código já existe");
     }
 
-    const product = new ProductEntity(productDto);
+    const product = new ProductEntity({ ...productDto, id: undefined, status: true });
     return this.productRepository.createProduct(
       product.codigo_produto,
       product.descricao_produto,
@@ -24,9 +24,14 @@ export class ProductService {
     return this.productRepository.getAllProducts();
   }
 
-  async updateProduct(id: string, data: Partial<CreateProductDTO>) {
+  async updateProduct(id: string, data: Partial<ProductDTO>) {
     const currentProduct = await this.productRepository.getProductById(id);
     const updatedProduct = new ProductEntity({ ...currentProduct, ...data });
+
+    const codeAlredyExists = await this.productRepository.getProductByCodigo(data.codigo_produto);
+    if (codeAlredyExists) {
+      throw new Error("Produto com o mesmo código já existe");
+    }
 
     if (currentProduct?.foto_produto && data.foto_produto && data.foto_produto !== currentProduct.foto_produto) {
       await deleteImageFromFirebase(currentProduct.foto_produto);
@@ -84,11 +89,10 @@ export class ProductService {
         const product = await this.productRepository.getProductById(id);
         
         if (product?.foto_produto) {
-            // Deleta a imagem do Firebase
             await deleteImageFromFirebase(product.foto_produto);
             
-            // Atualiza o produto removendo a referência da imagem
-            await this.productRepository.updateProduct(id, { foto_produto: null });
+            const updatedProduct = new ProductEntity({ ...product, foto_produto: null });
+            await this.productRepository.updateProduct(id, updatedProduct);
         }
     } catch (error) {
         throw error;
